requirements.txt
# Django and core dependencies
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Machine Learning and Data Processing
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2
nltk==3.8.1
vaderSentiment==3.3.2

# Data visualization
matplotlib==3.8.3
seaborn==0.13.1
plotly==5.17.0

# File processing
openpyxl==3.1.2
xlrd==2.0.1

# HTTP and API
requests==2.31.0
urllib3==2.1.0

# Environment and configuration
python-decouple==3.8
python-dotenv==1.0.0

# Development and testing
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Code quality
black==24.2.0
flake8==7.0.0
isort==5.13.2

# Security
cryptography==42.0.4
django-ratelimit==4.1.0

# Monitoring and logging
django-extensions==3.2.3
django-debug-toolbar==4.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.34

# Redis and Caching
django-redis==5.4.0

# Background Tasks
celery==5.3.4

# Error Tracking
sentry-sdk==1.38.0

# Frontend build tools (for development)
